pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Joanie <greystorm101@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.3.2"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
